+++ src/fife/video/sdl/sdlblendingfunctions.h
@@ -23,34 +23,10 @@ namespace FIFE
      * @param alpha Level of alphablending.
      * @param n Number of pixels.
      */
-    void SDL_BlendRow_RGBA8_to_RGBA8(uint8_t const * src, uint8_t* dst, uint32_t alpha, int32_t n);
-
-    /** Blends one row of n pixels from src with n pixels of dst.
-     *
-     * @param src Source.
-     * @param dst Destiny.
-     * @param alpha Level of alphablending.
-     * @param n Number of pixels.
-     */
-    void SDL_BlendRow_RGBA8_to_RGB8(uint8_t const * src, uint8_t* dst, uint32_t alpha, int32_t n);
-
-    /** Blends one row of n pixels from src with n pixels of dst.
-     *
-     * @param src Source.
-     * @param dst Destiny.
-     * @param alpha Level of alphablending.
-     * @param n Number of pixels.
-     */
-    void SDL_BlendRow_RGBA8_to_RGB565(uint8_t const * src, uint8_t* dst, uint32_t alpha, int32_t n);
-
-    /** Blends one row of n pixels from src with n pixels of dst.
-     *
-     * @param src Source.
-     * @param dst Destiny.
-     * @param alpha Level of alphablending.
-     * @param n Number of pixels.
-     */
-    void SDL_BlendRow_RGBA4_to_RGB565(uint8_t const * src, uint8_t* dst, uint32_t alpha, int32_t n);
+    void SDL_BlendRow_RGBA8_to_RGBA8(uint8_t const * src, uint8_t* dst, uint32_t alpha, size_t n);
+    void SDL_BlendRow_RGBA8_to_RGB8(uint8_t const * src, uint8_t* dst, uint32_t alpha, size_t n);
+    void SDL_BlendRow_RGBA8_to_RGB565(uint8_t const * src, uint8_t* dst, uint32_t alpha, size_t n);
+    void SDL_BlendRow_RGBA4_to_RGB565(uint8_t const * src, uint8_t* dst, uint32_t alpha, size_t n);
 
 } // namespace FIFE
